MEDIUM: mworker/cli: close child and parent fds, setup listeners
This commit is a part of series to add the support of 4 primary reload use-cases
for the new master-worker architecture.

Basically, this is the continuation of the previous commit. So, here after the
fork, worker process closes the "master" end of the copied CLI sockpair and
binds its end, ipc_fd[1], to the GLOBAL proxy listener.
mworker_cli_global_proxy_new_listener() guarantees that GLOBAL proxy will be
created, if it wasn't the case before.

Master process at first allocates the MASTER proxy in
mworker_create_master_cli() and then closes the "worker" end of the copied CLI
sockpair and binds its end, ipc_fd[0], to the created MASTER proxy.