MEDIUM: mworker: add support for case when new worker dies
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

The case, when the new worker fails while it parses its configuration or while
it tries to apply it, could be considered as the new one, because the master
process is no longer need to reexec again. The master simply keeps the previous
worker (forked before the reload) and it lets the new one to exit with failure.

So, at first to distinguish between the new worker and the previous one let's
add a new process state PROC_O_INIT and let's set it, when the memory is
allocated for the new worker in the processes list.

When the new worker exits, in the master process context
(mworker_catch_sigchld() scope) we need to stop a MASTER proxy listener and we
need to drop the server, attached to new worker's CLI sockpair (it's inherited
in master). Then we explicitly delete master's end of this sockpair
(child->ipc_fd[0]) from the fdtab and we free the memory allocated for the
worker process. on_new_child_failure() is called before the
clean up to signal systemd that reload/load was failed.

If the new worker fails during the first start, so there is no any previous
worker, master process also exits with worker's exitcode.