MINOR: startup: clean up master code
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

With this new architecture we no longer need to reexec master, if the new
forked worker fails during its initialization step. Basically in this case, we
just need to do some clean up in the master process context
(mworker_catch_sigchld), which was introduced in the previous commit. Master
process always stays in its polling loop, except cases when it receives a TERM
signal or it just crashes.

In case, when TERM signal is received by the master, it sends it to all workers
before exiting the polling loop. So, now no need for the master after exiting
its polling loop to set "proc_self" pointer to the newly forked worker. For the
moment we still keep the code, where the master loops the last time through
the processes list before exiting, closes its ends of inherited master CLI
sockpairs and frees all processes entries. But this part will be also
simplified in future.