MINOR: startup: split init
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

Let's split here the init() function into separate initialization routines,
which set and apply modes, write process PID in a pidfile, if provided, etc.

The big part of the former init(), which called functions to allocate pools,
to initialize proxies, to calculate maxconn and to do some post checks was just
encasulated as is into step_init_2(). It will take some time to figure out
exactly which parts of this initialization block are really necessary for the
master process and which ones it could skip. So, for the moment step_init_2()
is called for all runtime modes.