MINOR: startup: split main: add step_init_3
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

The first part of main(), just after calling the former init() and before
trying to bind listeners, need to be also encapsulated into a separate
step_init_3() as it is. It contains important blocks to register signals, to
apply memory and nofile limits, etc. The order of these blocks should be also
preserved (especially the signals part).

For the moment step_init_3() must be also executed for all runtime modes.