MINOR: startup: simplify check for calling sock_get_old_sockets
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

MODE_CHECK and MODE_CHECK_CONDITION are applied now very early in
step_init_1() and step_init_2() in order to check the configuration or to check
some condition provided via the command line. When these checks have
terminated, the main process exits. So, no longer need to verify these modes at
the moment, when the current process have already done its basic initialization
routines and is asking for listeners sockets from the previously started one.