MINOR: startup: encapsulate sock_get_old_sockets in a function
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

Let's encapsulate here the code, that calls sock_get_old_sockets() to obtain
listeners sockets from the previous process into a separate function. This
will make the code of main() more readable and we can move this new function
(if we might need so) in future.