MINOR: mworker: add run_master_in_recovery_mode
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

In this commit we add run_master_in_recovery_mode(), which groups all necessary
initialization steps, which master should perform to be able to enter in its
polling loop (run_master()), when it fails while parsing its new config.

As exit_on_failure() is now adapted for master recovery mode, let's register
it as atexit handler, when master enters in this mode. And let's remove
atexit_flag variable for master, because we no longer use it.