MEDIUM: mworker: add support of master recovery mode
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

Let's slightly refactor here read_cfg_in_discovery_mode() in order to call
run_master_in_recovery_mode() in cases, when master fails to load or to parse
its new configuration after relaod. Warning messages are mandatory before
calling the run_master_in_recovery_mode() as this allows to stop haproxy
with error, if it was launched in zero-warning mode.

So, in recovery mode, master does not launch any worker. It just performs its
necessary initialization routines and enters in its polling loop to continue
to monitor the existed worker process.