MINOR: startup: make progname global in haproxy.c
This is the first commit in a series to add a support of the 5-th reload
use case, when the master process fails to read its new configuration. In this
case it just need to perform its initialization steps and keep the existed
worker.

To add the support for this last use case we need to split init() and main()
in a shorter steps in order to encapsulate necessary initialization routines
into separate functions.

Let's make here progname as a global variable for haproxy.c, as it will be
used in error messages in these new initialization functions.