MINOR: mworker: remove atexit_flag for worker
This commit is a part of series to add a support of the 5-th reload
use case, when the master fails after reload, while loading or parsing
its new configuration.

When worker fails all clean up is done now in on_new_child_failure() and in
mworker_catch_sigchld(). We no longer register any atexit function for the
worker, so this variable becames redundant.