MINOR: mworker: add and set state PROC_O_INIT for new worker
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

Here, to distinguish between the new worker and the previous one let's add a
new process state PROC_O_INIT and let's set it, when the memory is allocated
for the new worker in the processes list.