MINOR: mworker/cli: create MASTER proxy before mcli listeners
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

For the master process we always need to create a MASTER proxy, even if
master cli settings were not provided via command line, because now we bind a
listener in the master process context at ipc_fd[0]. So, MASTER proxy should be
already allocated at this moment.