BUG/MINOR: mworker-prog: don't warn about deprecated section with expose-deprecated-directives
As master parses now expose-deprecated-directives option, let's emit warning
about deprecated 'progam' section only in case, if this option wasn't set in
the 'global' section. This allows to people, who can't change the
configuration immediately to continue to start the process in zero-warning
mode.

Adjust the warning message accordingly. As expose-deprecated-directives
option is a 'global' section keyword, this section must always precede
any 'program' section, if users still continue to keep 'program' section.

This doesn't need to be backported, as related to the latest changes in
the master-worker architecture.