BUG/MINOR: cfgparse-tcp: no need to set LSTCHK_SYSADM for bind namespace
When haproxy changes its uid from root to a regular user and checks then,
if all needed capabilities were preserved, listeners sockets are already
created and bound. So, there is no need to set LSTCHK_SYSADM in
bind_parse_namespace().

At contrast, "namespace" is also used for server settings for outgoing
connections. So, we still need to set LSTCHK_SYSADM, when we parse a server
line. Like this, after uid switch, we could check if we've preserved
CAP_SYS_ADMIN, in order to access sockets in server's namespace.