MINOR: cfgparse: detect empty lines within ACL and env variables
Emty lines in ACL expressions or conditions like below:

	http-request deny if { path_dir '' }
	acl bad_path path_sub '%2E' '' '%2F'

stop parse_line tokenizer, because it treats '\0' as the end of arguments list.

Users in these cases may experience some problems, like not very comprehensive
error messages in stderr (the first example) or requests with '%2F' in path are
still accepted.

The same problem with env variables, if their values are set explicitly as
empty lines, or not set at all:

	setenv MYVAR ""
	.if streq("$ENABLE_SSL",yes)
	...

	$ env | grep ENABLE_SSL
	ENABLE_SSL=

So, let's emit more comprehensive warnings in such cases about the detected
emty lines. This should help users to fix their configurations.

This resolves issue #2944. Not need to be backported.